Move all player information into data folder
